Extract editor validation modules
