+++ src/game_data_validation.c
@@ -132,7 +132,7 @@ static bool validation_key_name_valid(const char *name)
     return SDL_GetScancodeFromName(name) != SDL_SCANCODE_UNKNOWN;
 }
 
-static bool validation_mouse_button_name_valid(const char *name)
+bool validation_mouse_button_name_valid(const char *name)
 {
     return name != NULL &&
            (SDL_strcmp(name, "LEFT") == 0 || SDL_strcmp(name, "MIDDLE") == 0 || SDL_strcmp(name, "RIGHT") == 0 ||
@@ -499,8 +499,7 @@ bool require_ref(validation_context *ctx, const name_table *table, const char *k
     return true;
 }
 
-static bool require_actor_ref(validation_context *ctx, const validation_names *names, const char *name,
-                              const char *json_path)
+bool require_actor_ref(validation_context *ctx, const validation_names *names, const char *name, const char *json_path)
 {
     if (name == NULL || name[0] == '\0')
         return validation_error(ctx, json_path, "missing actor reference");
@@ -6319,355 +6318,6 @@ static bool validate_actor_archetypes_and_pools(validation_context *ctx, yyjson_
     return true;
 }
 
-static bool editor_string_field_valid(validation_context *ctx, yyjson_val *metadata, const char *key,
-                                      const char *json_path)
-{
-    yyjson_val *value = obj_get(metadata, key);
-    if (value == NULL)
-        return true;
-    if (!yyjson_is_str(value) || yyjson_get_str(value)[0] == '\0')
-        return validation_error(ctx, json_path, "editor %s must be a non-empty string", key);
-    return true;
-}
-
-static bool editor_tags_valid(validation_context *ctx, yyjson_val *tags, const char *json_path)
-{
-    if (tags == NULL)
-        return true;
-    if (!yyjson_is_arr(tags))
-        return validation_error(ctx, json_path, "editor tags must be an array");
-    for (size_t i = 0; i < yyjson_arr_size(tags); ++i)
-    {
-        char path[PATH_BUFFER_SIZE];
-        format_path(path, sizeof(path), "%s[%zu]", json_path, i);
-        yyjson_val *tag = yyjson_arr_get(tags, i);
-        if (!yyjson_is_str(tag) || yyjson_get_str(tag)[0] == '\0')
-            return validation_error(ctx, path, "editor tag entries must be non-empty strings");
-    }
-    return true;
-}
-
-static bool editor_vec_positive(yyjson_val *value)
-{
-    if (!is_vec_array(value, 3))
-        return false;
-    for (size_t i = 0; i < yyjson_arr_size(value); ++i)
-    {
-        if (yyjson_get_num(yyjson_arr_get(value, i)) <= 0.0)
-            return false;
-    }
-    return true;
-}
-
-static bool editor_preview_primitive_valid(const char *primitive)
-{
-    static const char *const valid[] = {"none",  "cube",  "sphere", "capsule", "sprite",
-                                        "model", "light", "volume", "sector"};
-    for (size_t i = 0; primitive != NULL && i < SDL_arraysize(valid); ++i)
-    {
-        if (SDL_strcmp(primitive, valid[i]) == 0)
-            return true;
-    }
-    return false;
-}
-
-static bool editor_exposed_property_type_valid(const char *type)
-{
-    static const char *const valid[] = {"bool", "int", "float", "string", "vec2", "vec3", "color", "enum"};
-    for (size_t i = 0; type != NULL && i < SDL_arraysize(valid); ++i)
-    {
-        if (SDL_strcmp(type, valid[i]) == 0)
-            return true;
-    }
-    return false;
-}
-
-static bool editor_exposed_property_default_valid(yyjson_val *value, const char *type)
-{
-    if (value == NULL)
-        return true;
-    if (type == NULL)
-    {
-        return yyjson_is_str(value) || yyjson_is_int(value) || yyjson_is_real(value) || yyjson_is_bool(value) ||
-               is_exact_vec_array(value, 2) || is_exact_vec_array(value, 3) || is_exact_vec_array(value, 4);
-    }
-    if (SDL_strcmp(type, "bool") == 0)
-        return yyjson_is_bool(value);
-    if (SDL_strcmp(type, "int") == 0)
-        return yyjson_is_int(value);
-    if (SDL_strcmp(type, "float") == 0)
-        return yyjson_is_num(value);
-    if (SDL_strcmp(type, "string") == 0 || SDL_strcmp(type, "enum") == 0)
-        return yyjson_is_str(value);
-    if (SDL_strcmp(type, "vec2") == 0)
-        return is_exact_vec_array(value, 2);
-    if (SDL_strcmp(type, "vec3") == 0)
-        return is_exact_vec_array(value, 3);
-    if (SDL_strcmp(type, "color") == 0)
-        return is_exact_vec_array(value, 3) || is_exact_vec_array(value, 4);
-    return false;
-}
-
-static bool validate_editor_preview(validation_context *ctx, yyjson_val *preview, const char *json_path)
-{
-    if (preview == NULL)
-        return true;
-    if (!yyjson_is_obj(preview))
-        return validation_error(ctx, json_path, "editor preview must be an object");
-
-    const char *primitive = json_string(preview, "primitive");
-    if (primitive != NULL && !editor_preview_primitive_valid(primitive))
-        return validation_error(ctx, json_path, "editor preview primitive is unknown");
-    yyjson_val *asset = obj_get(preview, "asset");
-    if (asset != NULL && (!yyjson_is_str(asset) || yyjson_get_str(asset)[0] == '\0'))
-        return validation_error(ctx, json_path, "editor preview asset must be a non-empty string");
-    yyjson_val *color = obj_get(preview, "color");
-    if (color != NULL && !is_vec_array(color, 3))
-        return validation_error(ctx, json_path, "editor preview color must be a vec3 or vec4");
-    return true;
-}
-
-static bool validate_editor_bounds(validation_context *ctx, yyjson_val *bounds, const char *json_path)
-{
-    if (bounds == NULL)
-        return true;
-    if (!yyjson_is_obj(bounds))
-        return validation_error(ctx, json_path, "editor bounds must be an object");
-    yyjson_val *center = obj_get(bounds, "center");
-    if (center != NULL && !is_vec_array(center, 3))
-        return validation_error(ctx, json_path, "editor bounds center must be a vec3");
-    yyjson_val *size = obj_get(bounds, "size");
-    if (size != NULL && !editor_vec_positive(size))
-        return validation_error(ctx, json_path, "editor bounds size must be a positive vec3");
-    yyjson_val *half_extents = obj_get(bounds, "half_extents");
-    if (half_extents != NULL && !editor_vec_positive(half_extents))
-        return validation_error(ctx, json_path, "editor bounds half_extents must be a positive vec3");
-    yyjson_val *radius = obj_get(bounds, "radius");
-    if (radius != NULL && (!yyjson_is_num(radius) || yyjson_get_num(radius) <= 0.0))
-        return validation_error(ctx, json_path, "editor bounds radius must be positive");
-    return true;
-}
-
-static bool validate_editor_snap(validation_context *ctx, yyjson_val *snap, const char *json_path)
-{
-    if (snap == NULL)
-        return true;
-    if (!yyjson_is_obj(snap))
-        return validation_error(ctx, json_path, "editor snap must be an object");
-    yyjson_val *grid = obj_get(snap, "grid");
-    if (grid != NULL)
-    {
-        if (yyjson_is_num(grid))
-        {
-            if (yyjson_get_num(grid) <= 0.0)
-                return validation_error(ctx, json_path, "editor snap grid must be positive");
-        }
-        else if (!editor_vec_positive(grid))
-        {
-            return validation_error(ctx, json_path, "editor snap grid must be a positive number or vec3");
-        }
-    }
-    yyjson_val *rotation = obj_get(snap, "rotation_degrees");
-    if (rotation != NULL && (!yyjson_is_num(rotation) || yyjson_get_num(rotation) <= 0.0))
-        return validation_error(ctx, json_path, "editor snap rotation_degrees must be positive");
-    yyjson_val *align = obj_get(snap, "align_to_floor");
-    if (align != NULL && !yyjson_is_bool(align))
-        return validation_error(ctx, json_path, "editor snap align_to_floor must be a boolean");
-    return true;
-}
-
-static bool validate_editor_exposed_properties(validation_context *ctx, yyjson_val *properties, const char *json_path)
-{
-    if (properties == NULL)
-        return true;
-    if (!yyjson_is_arr(properties))
-        return validation_error(ctx, json_path, "editor exposed_properties must be an array");
-    name_table names;
-    SDL_zero(names);
-    bool ok = true;
-    for (size_t i = 0; ok && i < yyjson_arr_size(properties); ++i)
-    {
-        char path[PATH_BUFFER_SIZE];
-        format_path(path, sizeof(path), "%s[%zu]", json_path, i);
-        yyjson_val *property = yyjson_arr_get(properties, i);
-        if (!yyjson_is_obj(property))
-        {
-            ok = validation_error(ctx, path, "editor exposed property entries must be objects");
-            break;
-        }
-        ok = require_unique_name(ctx, &names, "editor exposed property", json_string(property, "name"), path) &&
-             editor_string_field_valid(ctx, property, "display_name", path) &&
-             editor_string_field_valid(ctx, property, "description", path);
-        const char *type = json_string(property, "type");
-        if (ok && type != NULL && !editor_exposed_property_type_valid(type))
-            ok = validation_error(ctx, path, "editor exposed property type is unknown");
-        yyjson_val *min_value = obj_get(property, "min");
-        yyjson_val *max_value = obj_get(property, "max");
-        if (ok &&
-            ((min_value != NULL && !yyjson_is_num(min_value)) || (max_value != NULL && !yyjson_is_num(max_value))))
-            ok = validation_error(ctx, path, "editor exposed property min and max must be numbers");
-        yyjson_val *default_value = obj_get(property, "default");
-        if (ok && !editor_exposed_property_default_valid(default_value, type))
-            ok = validation_error(ctx, path, "editor exposed property default does not match its type");
-    }
-    name_table_destroy(&names);
-    return ok;
-}
-
-bool validate_editor_metadata(validation_context *ctx, yyjson_val *metadata, const char *json_path,
-                              validation_names *names, bool allow_templates)
-{
-    if (metadata == NULL)
-        return true;
-    if (!yyjson_is_obj(metadata))
-        return validation_error(ctx, json_path, "editor metadata must be an object");
-
-    if (!editor_string_field_valid(ctx, metadata, "display_name", json_path) ||
-        !editor_string_field_valid(ctx, metadata, "stable_id", json_path) ||
-        !editor_string_field_valid(ctx, metadata, "description", json_path) ||
-        !editor_string_field_valid(ctx, metadata, "category", json_path) ||
-        !editor_string_field_valid(ctx, metadata, "group", json_path) ||
-        !editor_string_field_valid(ctx, metadata, "prefab", json_path) ||
-        !editor_string_field_valid(ctx, metadata, "archetype", json_path) ||
-        !editor_string_field_valid(ctx, metadata, "icon", json_path) ||
-        !editor_string_field_valid(ctx, metadata, "preview_asset", json_path) ||
-        !editor_tags_valid(ctx, obj_get(metadata, "tags"), json_path) ||
-        !validate_editor_preview(ctx, obj_get(metadata, "preview"), json_path) ||
-        !validate_editor_bounds(ctx, obj_get(metadata, "bounds"), json_path) ||
-        !validate_editor_snap(ctx, obj_get(metadata, "snap"), json_path) ||
-        !validate_editor_exposed_properties(ctx, obj_get(metadata, "exposed_properties"), json_path))
-    {
-        return false;
-    }
-
-    const char *test_scene = json_string(metadata, "test_scene");
-    if (test_scene != NULL && !require_ref(ctx, &names->scenes, "scene", test_scene, json_path))
-        return false;
-
-    yyjson_val *templates = obj_get(metadata, "templates");
-    if (templates == NULL)
-        return true;
-    if (!allow_templates)
-        return validation_error(ctx, json_path, "editor templates are only allowed on root editor metadata");
-    if (!yyjson_is_arr(templates))
-        return validation_error(ctx, json_path, "editor templates must be an array");
-
-    name_table template_names;
-    SDL_zero(template_names);
-    bool ok = true;
-    for (size_t i = 0; ok && i < yyjson_arr_size(templates); ++i)
-    {
-        char path[PATH_BUFFER_SIZE];
-        format_path(path, sizeof(path), "%s.templates[%zu]", json_path, i);
-        yyjson_val *entry = yyjson_arr_get(templates, i);
-        if (!yyjson_is_obj(entry))
-        {
-            ok = validation_error(ctx, path, "editor template entries must be objects");
-            break;
-        }
-        ok = require_unique_name(ctx, &template_names, "editor template", json_string(entry, "name"), path) &&
-             editor_string_field_valid(ctx, entry, "source", path) &&
-             editor_string_field_valid(ctx, entry, "source_kind", path) &&
-             validate_editor_metadata(ctx, entry, path, names, false);
-    }
-    name_table_destroy(&template_names);
-    return ok;
-}
-
-static bool validate_editor_metadata_tree(validation_context *ctx, yyjson_val *root, validation_names *names)
-{
-    if (!validate_editor_metadata(ctx, obj_get(root, "editor"), "$.editor", names, true))
-        return false;
-
-    yyjson_val *entities = obj_get(root, "entities");
-    for (size_t i = 0; yyjson_is_arr(entities) && i < yyjson_arr_size(entities); ++i)
-    {
-        char path[PATH_BUFFER_SIZE];
-        format_path(path, sizeof(path), "$.entities[%zu].editor", i);
-        if (!validate_editor_metadata(ctx, obj_get(yyjson_arr_get(entities, i), "editor"), path, names, false))
-            return false;
-    }
-
-    yyjson_val *archetypes = obj_get(root, "actor_archetypes");
-    for (size_t i = 0; yyjson_is_arr(archetypes) && i < yyjson_arr_size(archetypes); ++i)
-    {
-        char path[PATH_BUFFER_SIZE];
-        format_path(path, sizeof(path), "$.actor_archetypes[%zu].editor", i);
-        if (!validate_editor_metadata(ctx, obj_get(yyjson_arr_get(archetypes, i), "editor"), path, names, false))
-            return false;
-    }
-
-    yyjson_val *instances = obj_get(root, "actor_instances");
-    for (size_t i = 0; yyjson_is_arr(instances) && i < yyjson_arr_size(instances); ++i)
-    {
-        char path[PATH_BUFFER_SIZE];
-        format_path(path, sizeof(path), "$.actor_instances[%zu].editor", i);
-        if (!validate_editor_metadata(ctx, obj_get(yyjson_arr_get(instances, i), "editor"), path, names, false))
-            return false;
-    }
-
-    yyjson_val *pools = obj_get(root, "actor_pools");
-    for (size_t i = 0; yyjson_is_arr(pools) && i < yyjson_arr_size(pools); ++i)
-    {
-        char path[PATH_BUFFER_SIZE];
-        format_path(path, sizeof(path), "$.actor_pools[%zu].editor", i);
-        if (!validate_editor_metadata(ctx, obj_get(yyjson_arr_get(pools, i), "editor"), path, names, false))
-            return false;
-    }
-    return true;
-}
-
-static bool validate_editor_player_starts(validation_context *ctx, yyjson_val *root, validation_names *names)
-{
-    yyjson_val *starts = obj_get(root, "editor_player_starts");
-    if (starts == NULL)
-        return true;
-    if (!yyjson_is_arr(starts))
-        return validation_error(ctx, "$.editor_player_starts", "editor_player_starts must be an array");
-
-    name_table start_names;
-    SDL_zero(start_names);
-    bool ok = true;
-    for (size_t i = 0; ok && i < yyjson_arr_size(starts); ++i)
-    {
-        char path[PATH_BUFFER_SIZE];
-        format_path(path, sizeof(path), "$.editor_player_starts[%zu]", i);
-        yyjson_val *start = yyjson_arr_get(starts, i);
-        if (!yyjson_is_obj(start))
-        {
-            ok = validation_error(ctx, path, "editor player start entries must be objects");
-            break;
-        }
-        const char *scene = json_string(start, "scene");
-        const char *target = json_string(start, "target");
-        yyjson_val *yaw = obj_get(start, "yaw");
-        yyjson_val *pitch = obj_get(start, "pitch");
-        ok = require_unique_name(ctx, &start_names, "editor player start", json_string(start, "name"), path) &&
-             (scene == NULL || require_ref(ctx, &names->scenes, "scene", scene, path)) &&
-             (target == NULL || require_actor_ref(ctx, names, target, path)) &&
-             is_exact_vec_array(obj_get(start, "position"), 3) && (yaw == NULL || yyjson_is_num(yaw)) &&
-             (pitch == NULL || yyjson_is_num(pitch));
-        if (ok)
-            continue;
-        if (!ctx->failed)
-            ok = validation_error(ctx, path, "editor player start requires position vec3 and numeric yaw/pitch");
-    }
-    name_table_destroy(&start_names);
-    return ok;
-}
-
-bool require_unique_editor_stable_id(validation_context *ctx, name_table *stable_ids, yyjson_val *json,
-                                     const char *json_path)
-{
-    yyjson_val *editor = obj_get(json, "editor");
-    const char *stable_id = json_string(editor, "stable_id");
-    if (stable_id == NULL)
-        return true;
-    char path[PATH_BUFFER_SIZE];
-    format_path(path, sizeof(path), "%s.editor", json_path);
-    return require_unique_name(ctx, stable_ids, "editor stable id", stable_id, path);
-}
-
 static bool is_tween_easing(const char *easing)
 {
     return easing == NULL || SDL_strcmp(easing, "linear") == 0 || SDL_strcmp(easing, "in_quad") == 0 ||
@@ -10803,22 +10453,6 @@ static bool validate_scene_skybox(validation_context *ctx, yyjson_val *scene_roo
     return true;
 }
 
-static bool editor_model_filter_name_valid(const char *value)
-{
-    return value != NULL && (SDL_strcmp(value, "all") == 0 || SDL_strcmp(value, "sector_levels") == 0 ||
-                             SDL_strcmp(value, "sector") == 0 || SDL_strcmp(value, "brush_worlds") == 0 ||
-                             SDL_strcmp(value, "brush") == 0);
-}
-
-static bool editor_debug_flag_name_valid(const char *value)
-{
-    return value != NULL && (SDL_strcmp(value, "all") == 0 || SDL_strcmp(value, "world_bounds") == 0 ||
-                             SDL_strcmp(value, "selection_bounds") == 0 || SDL_strcmp(value, "trace_ray") == 0 ||
-                             SDL_strcmp(value, "face_normal") == 0 || SDL_strcmp(value, "hit_marker") == 0 ||
-                             SDL_strcmp(value, "command_preview") == 0 || SDL_strcmp(value, "work_plane_grid") == 0 ||
-                             SDL_strcmp(value, "grid") == 0);
-}
-
 static bool editor_command_name_valid(const char *value)
 {
     return value != NULL &&
@@ -10831,426 +10465,6 @@ static bool editor_command_target_name_valid(const char *value)
     return value != NULL &&
            (SDL_strcmp(value, "selection") == 0 || SDL_strcmp(value, "world") == 0 ||
             SDL_strcmp(value, "element") == 0 || SDL_strcmp(value, "face") == 0 || SDL_strcmp(value, "material") == 0);
-}
-
-static bool validate_string_or_string_array_names(validation_context *ctx, yyjson_val *value, const char *path,
-                                                  const char *label, bool (*name_valid)(const char *name))
-{
-    if (value == NULL)
-        return true;
-    if (yyjson_is_str(value))
-        return name_valid(yyjson_get_str(value))
-                   ? true
-                   : validation_error(ctx, path, "unsupported %s '%s'", label, yyjson_get_str(value));
-    if (!yyjson_is_arr(value))
-        return validation_error(ctx, path, "%s must be a string or string array", label);
-    for (size_t i = 0; i < yyjson_arr_size(value); ++i)
-    {
-        char entry_path[PATH_BUFFER_SIZE];
-        format_path(entry_path, sizeof(entry_path), "%s[%zu]", path, i);
-        yyjson_val *entry = yyjson_arr_get(value, i);
-        if (!yyjson_is_str(entry) || !name_valid(yyjson_get_str(entry)))
-            return validation_error(ctx, entry_path, "unsupported %s '%s'", label,
-                                    yyjson_is_str(entry) ? yyjson_get_str(entry) : "<non-string>");
-    }
-    return true;
-}
-
-static bool validate_scene_editor_outputs(validation_context *ctx, yyjson_val *outputs, const char *json_path)
-{
-    if (outputs == NULL)
-        return true;
-    if (!yyjson_is_obj(outputs))
-        return validation_error(ctx, json_path, "scene editor selection outputs must be an object");
-
-    yyjson_val *key;
-    yyjson_obj_iter iter;
-    yyjson_obj_iter_init(outputs, &iter);
-    while ((key = yyjson_obj_iter_next(&iter)) != NULL)
-    {
-        yyjson_val *value = yyjson_obj_iter_get_val(key);
-        if (!yyjson_is_str(value) || yyjson_get_str(value)[0] == '\0')
-            return validation_error(ctx, json_path, "scene editor selection output values must be non-empty strings");
-    }
-    return true;
-}
-
-static bool validate_optional_non_empty_string(validation_context *ctx, yyjson_val *value, const char *json_path,
-                                               const char *description)
-{
-    if (value == NULL)
-        return true;
-    if (!yyjson_is_str(value) || yyjson_get_str(value)[0] == '\0')
-        return validation_error(ctx, json_path, "%s must be a non-empty string", description);
-    return true;
-}
-
-static bool validate_optional_number(validation_context *ctx, yyjson_val *value, const char *json_path,
-                                     const char *description)
-{
-    if (value == NULL)
-        return true;
-    if (!yyjson_is_num(value))
-        return validation_error(ctx, json_path, "%s must be a number", description);
-    return true;
-}
-
-static bool validate_optional_positive_number(validation_context *ctx, yyjson_val *value, const char *json_path,
-                                              const char *description)
-{
-    if (value == NULL)
-        return true;
-    if (!yyjson_is_num(value) || yyjson_get_num(value) <= 0.0)
-        return validation_error(ctx, json_path, "%s must be positive", description);
-    return true;
-}
-
-static bool validate_optional_non_negative_number(validation_context *ctx, yyjson_val *value, const char *json_path,
-                                                  const char *description)
-{
-    if (value == NULL)
-        return true;
-    if (!yyjson_is_num(value) || yyjson_get_num(value) < 0.0)
-        return validation_error(ctx, json_path, "%s must be non-negative", description);
-    return true;
-}
-
-static bool validate_scene_editor_camera_screen_trace(validation_context *ctx, yyjson_val *trace,
-                                                      const char *trace_path, validation_names *names)
-{
-    yyjson_val *camera_value = obj_get(trace, "camera");
-    if (!validate_optional_non_empty_string(ctx, camera_value, trace_path, "scene editor camera_screen trace camera"))
-        return false;
-    const char *camera = json_string(trace, "camera");
-    if (camera != NULL && !require_ref(ctx, &names->cameras, "camera", camera, trace_path))
-        return false;
-
-    yyjson_val *screen = obj_get(trace, "screen");
-    if (screen != NULL && !is_exact_vec_array(screen, 2))
-        return validation_error(ctx, trace_path, "scene editor camera_screen trace screen must be a vec2");
-    yyjson_val *viewport = obj_get(trace, "viewport");
-    if (viewport != NULL &&
-        (!is_exact_vec_array(viewport, 2) || !numeric_array_values_in_range(viewport, 0.000001, DBL_MAX)))
-        return validation_error(ctx, trace_path, "scene editor camera_screen trace viewport must be a positive vec2");
-
-    char field_path[PATH_BUFFER_SIZE];
-    static const char *const numeric_keys[] = {"screen_x", "screen_y"};
-    for (size_t i = 0; i < SDL_arraysize(numeric_keys); ++i)
-    {
-        format_path(field_path, sizeof(field_path), "%s.%s", trace_path, numeric_keys[i]);
-        if (!validate_optional_number(ctx, obj_get(trace, numeric_keys[i]), field_path,
-                                      "scene editor camera_screen trace coordinate"))
-            return false;
-    }
-
-    static const char *const positive_keys[] = {"viewport_width", "viewport_height", "far"};
-    for (size_t i = 0; i < SDL_arraysize(positive_keys); ++i)
-    {
-        format_path(field_path, sizeof(field_path), "%s.%s", trace_path, positive_keys[i]);
-        if (!validate_optional_positive_number(ctx, obj_get(trace, positive_keys[i]), field_path,
-                                               "scene editor camera_screen trace value"))
-            return false;
-    }
-    format_path(field_path, sizeof(field_path), "%s.near", trace_path);
-    if (!validate_optional_non_negative_number(ctx, obj_get(trace, "near"), field_path,
-                                               "scene editor camera_screen trace near"))
-        return false;
-
-    static const char *const string_keys[] = {"screen_x_key", "screen_y_key", "viewport_width_key",
-                                              "viewport_height_key"};
-    for (size_t i = 0; i < SDL_arraysize(string_keys); ++i)
-    {
-        format_path(field_path, sizeof(field_path), "%s.%s", trace_path, string_keys[i]);
-        if (!validate_optional_non_empty_string(ctx, obj_get(trace, string_keys[i]), field_path,
-                                                "scene editor camera_screen trace key"))
-            return false;
-    }
-
-    yyjson_val *near_value = obj_get(trace, "near");
-    yyjson_val *far_value = obj_get(trace, "far");
-    if (near_value != NULL && far_value != NULL && yyjson_get_num(far_value) <= yyjson_get_num(near_value))
-        return validation_error(ctx, trace_path, "scene editor camera_screen trace far must be greater than near");
-    return true;
-}
-
-static bool validate_scene_editor_work_plane(validation_context *ctx, yyjson_val *trace, const char *trace_path)
-{
-    yyjson_val *work_plane = obj_get(trace, "work_plane");
-    if (work_plane == NULL)
-        return true;
-    char work_plane_path[PATH_BUFFER_SIZE];
-    format_path(work_plane_path, sizeof(work_plane_path), "%s.work_plane", trace_path);
-    if (!yyjson_is_obj(work_plane))
-        return validation_error(ctx, work_plane_path, "scene editor selection work_plane must be an object");
-
-    yyjson_val *enabled = obj_get(work_plane, "enabled");
-    if (enabled != NULL && !yyjson_is_bool(enabled))
-        return validation_error(ctx, work_plane_path, "scene editor selection work_plane enabled must be a boolean");
-
-    yyjson_val *normal = obj_get(work_plane, "normal");
-    if (normal != NULL)
-    {
-        if (!is_exact_vec_array(normal, 3))
-            return validation_error(ctx, work_plane_path, "scene editor selection work_plane normal must be a vec3");
-        const double x = yyjson_get_num(yyjson_arr_get(normal, 0));
-        const double y = yyjson_get_num(yyjson_arr_get(normal, 1));
-        const double z = yyjson_get_num(yyjson_arr_get(normal, 2));
-        if ((x * x + y * y + z * z) <= 0.000001)
-            return validation_error(ctx, work_plane_path, "scene editor selection work_plane normal must be non-zero");
-    }
-
-    char distance_path[PATH_BUFFER_SIZE];
-    format_path(distance_path, sizeof(distance_path), "%s.distance", work_plane_path);
-    if (!validate_optional_number(ctx, obj_get(work_plane, "distance"), distance_path,
-                                  "scene editor selection work_plane distance"))
-    {
-        return false;
-    }
-    return true;
-}
-
-static bool validate_scene_editor_trace(validation_context *ctx, yyjson_val *trace, const char *trace_path,
-                                        validation_names *names)
-{
-    yyjson_val *source_value = obj_get(trace, "source");
-    if (!validate_optional_non_empty_string(ctx, source_value, trace_path, "scene editor selection trace source"))
-        return false;
-    const char *source = source_value != NULL ? yyjson_get_str(source_value) : "world";
-    if (SDL_strcmp(source, "world") == 0)
-    {
-        if (!is_exact_vec_array(obj_get(trace, "start"), 3))
-            return validation_error(ctx, trace_path, "scene editor selection trace requires a start vec3");
-        if (!is_exact_vec_array(obj_get(trace, "end"), 3))
-            return validation_error(ctx, trace_path, "scene editor selection trace requires an end vec3");
-    }
-    else if (SDL_strcmp(source, "camera_screen") == 0)
-    {
-        if (!validate_scene_editor_camera_screen_trace(ctx, trace, trace_path, names))
-            return false;
-    }
-    else
-    {
-        return validation_error(ctx, trace_path,
-                                "scene editor selection trace source must be 'world' or 'camera_screen'");
-    }
-
-    char contents_path[PATH_BUFFER_SIZE];
-    format_path(contents_path, sizeof(contents_path), "%s.contents_mask", trace_path);
-    if (!validate_brush_string_or_string_array(ctx, obj_get(trace, "contents_mask"), contents_path, "brush content",
-                                               brush_content_name_valid, false))
-        return false;
-    char filter_path[PATH_BUFFER_SIZE];
-    format_path(filter_path, sizeof(filter_path), "%s.model_filter", trace_path);
-    if (!validate_string_or_string_array_names(ctx, obj_get(trace, "model_filter"), filter_path, "world model filter",
-                                               editor_model_filter_name_valid))
-        return false;
-    if (!validate_scene_editor_work_plane(ctx, trace, trace_path))
-        return false;
-    return true;
-}
-
-static bool validate_scene_editor_selection_options(validation_context *ctx, yyjson_val *selection,
-                                                    const char *selection_path)
-{
-    yyjson_val *mode = obj_get(selection, "mode");
-    if (mode != NULL)
-    {
-        if (!yyjson_is_str(mode) || yyjson_get_str(mode)[0] == '\0')
-            return validation_error(ctx, selection_path, "scene editor selection mode must be a non-empty string");
-        const char *mode_name = yyjson_get_str(mode);
-        if (SDL_strcmp(mode_name, "hover") != 0 && SDL_strcmp(mode_name, "click") != 0)
-            return validation_error(ctx, selection_path, "scene editor selection mode must be 'hover' or 'click'");
-    }
-
-    yyjson_val *select_button = obj_get(selection, "select_button");
-    if (select_button != NULL)
-    {
-        if (!yyjson_is_str(select_button) || !validation_mouse_button_name_valid(yyjson_get_str(select_button)))
-            return validation_error(ctx, selection_path, "scene editor selection select_button must be a mouse button");
-    }
-
-    yyjson_val *clear_on_miss = obj_get(selection, "clear_on_miss");
-    if (clear_on_miss != NULL && !yyjson_is_bool(clear_on_miss))
-        return validation_error(ctx, selection_path, "scene editor selection clear_on_miss must be a boolean");
-    return true;
-}
-
-static bool validate_scene_editor_placement(validation_context *ctx, yyjson_val *placement, const char *placement_path,
-                                            validation_names *names)
-{
-    if (placement == NULL)
-        return true;
-    if (!yyjson_is_obj(placement))
-        return validation_error(ctx, placement_path, "scene editor placement must be an object");
-
-    const char *string_fields[] = {"tool_key", "snap_key", "default_tool"};
-    for (size_t i = 0; i < SDL_arraysize(string_fields); ++i)
-    {
-        char field_path[PATH_BUFFER_SIZE];
-        format_path(field_path, sizeof(field_path), "%s.%s", placement_path, string_fields[i]);
-        if (!validate_optional_non_empty_string(ctx, obj_get(placement, string_fields[i]), field_path,
-                                                "scene editor placement field"))
-            return false;
-    }
-    yyjson_val *default_snap = obj_get(placement, "default_snap");
-    if (default_snap != NULL && (!yyjson_is_num(default_snap) || yyjson_get_num(default_snap) <= 0.0))
-        return validation_error(ctx, placement_path, "scene editor placement default_snap must be positive");
-
-    yyjson_val *outputs = obj_get(placement, "outputs");
-    if (outputs != NULL && !yyjson_is_obj(outputs))
-        return validation_error(ctx, placement_path, "scene editor placement outputs must be an object");
-    static const char *const output_keys[] = {"active_key", "valid_key", "mode_key",       "kind_key",
-                                              "axis_key",   "world_key", "material_key",   "message_key",
-                                              "anchor_key", "snap_key",  "bounds_min_key", "bounds_max_key"};
-    for (size_t i = 0; yyjson_is_obj(outputs) && i < SDL_arraysize(output_keys); ++i)
-    {
-        yyjson_val *output = obj_get(outputs, output_keys[i]);
-        if (output != NULL && (!yyjson_is_str(output) || yyjson_get_str(output)[0] == '\0'))
-            return validation_error(ctx, placement_path,
-                                    "scene editor placement output keys must be non-empty strings");
-    }
-
-    yyjson_val *previews = obj_get(placement, "previews");
-    if (!yyjson_is_arr(previews) || yyjson_arr_size(previews) == 0)
-        return validation_error(ctx, placement_path, "scene editor placement previews must be a non-empty array");
-    for (size_t i = 0; i < yyjson_arr_size(previews); ++i)
-    {
-        char preview_path[PATH_BUFFER_SIZE];
-        format_path(preview_path, sizeof(preview_path), "%s.previews[%zu]", placement_path, i);
-        yyjson_val *preview = yyjson_arr_get(previews, i);
-        if (!yyjson_is_obj(preview))
-            return validation_error(ctx, preview_path, "scene editor placement preview must be an object");
-        if (!is_non_empty_string(preview, "mode"))
-            return validation_error(ctx, preview_path, "scene editor placement preview requires a non-empty mode");
-        const char *kind = json_string(preview, "kind");
-        if (kind == NULL)
-            kind = "box";
-        if (SDL_strcmp(kind, "box") != 0 && SDL_strcmp(kind, "player_start") != 0)
-            return validation_error(ctx, preview_path,
-                                    "scene editor placement preview kind must be box or player_start");
-        yyjson_val *axis_key = obj_get(preview, "axis_key");
-        if (axis_key != NULL && (!yyjson_is_str(axis_key) || yyjson_get_str(axis_key)[0] == '\0'))
-            return validation_error(ctx, preview_path, "scene editor placement preview axis_key must be non-empty");
-        const char *axis = json_string(preview, "axis");
-        if (axis != NULL && SDL_strcmp(axis, "x") != 0 && SDL_strcmp(axis, "z") != 0)
-            return validation_error(ctx, preview_path, "scene editor placement preview axis must be x or z");
-        yyjson_val *offset = obj_get(preview, "position_offset");
-        if (offset != NULL && !is_exact_vec_array(offset, 3))
-            return validation_error(ctx, preview_path, "scene editor placement preview position_offset must be a vec3");
-        yyjson_val *snap = obj_get(preview, "snap");
-        if (snap != NULL && (!yyjson_is_num(snap) || yyjson_get_num(snap) <= 0.0))
-            return validation_error(ctx, preview_path, "scene editor placement preview snap must be positive");
-        if (SDL_strcmp(kind, "player_start") == 0)
-        {
-            yyjson_val *size = obj_get(preview, "size");
-            if (size != NULL &&
-                (!is_exact_vec_array(size, 3) || yyjson_get_num(yyjson_arr_get(size, 0)) <= 0.0 ||
-                 yyjson_get_num(yyjson_arr_get(size, 1)) <= 0.0 || yyjson_get_num(yyjson_arr_get(size, 2)) <= 0.0))
-            {
-                return validation_error(ctx, preview_path,
-                                        "scene editor placement player_start size must be a positive vec3");
-            }
-            continue;
-        }
-        if (!require_ref(ctx, &names->brush_worlds, "brush world", json_string(preview, "world"), preview_path))
-            return false;
-        if (!is_non_empty_string(preview, "material"))
-            return validation_error(ctx, preview_path, "scene editor placement box preview requires a material");
-        yyjson_val *min = obj_get(preview, "min");
-        yyjson_val *max = obj_get(preview, "max");
-        if (!is_exact_vec_array(min, 3) || !is_exact_vec_array(max, 3))
-            return validation_error(ctx, preview_path, "scene editor placement box preview requires min and max vec3");
-        if (!(yyjson_get_num(yyjson_arr_get(min, 0)) < yyjson_get_num(yyjson_arr_get(max, 0)) &&
-              yyjson_get_num(yyjson_arr_get(min, 1)) < yyjson_get_num(yyjson_arr_get(max, 1)) &&
-              yyjson_get_num(yyjson_arr_get(min, 2)) < yyjson_get_num(yyjson_arr_get(max, 2))))
-        {
-            return validation_error(ctx, preview_path, "scene editor placement box preview bounds require min < max");
-        }
-    }
-    return true;
-}
-
-static bool validate_scene_editor_tooling(validation_context *ctx, yyjson_val *scene_root, const char *json_path,
-                                          validation_names *names)
-{
-    yyjson_val *editor = obj_get(scene_root, "editor");
-    if (editor == NULL)
-        return true;
-    if (!yyjson_is_obj(editor))
-        return validation_error(ctx, json_path, "scene editor must be an object");
-
-    yyjson_val *selection = obj_get(editor, "selection");
-    if (selection != NULL)
-    {
-        char selection_path[PATH_BUFFER_SIZE];
-        format_path(selection_path, sizeof(selection_path), "%s.editor.selection", json_path);
-        if (!yyjson_is_obj(selection))
-            return validation_error(ctx, selection_path, "scene editor selection must be an object");
-        if (!validate_scene_editor_selection_options(ctx, selection, selection_path))
-            return false;
-        yyjson_val *trace = obj_get(selection, "trace");
-        if (!yyjson_is_obj(trace))
-            return validation_error(ctx, selection_path, "scene editor selection requires a trace object");
-        char trace_path[PATH_BUFFER_SIZE];
-        format_path(trace_path, sizeof(trace_path), "%s.trace", selection_path);
-        if (!validate_scene_editor_trace(ctx, trace, trace_path, names))
-            return false;
-        char outputs_path[PATH_BUFFER_SIZE];
-        format_path(outputs_path, sizeof(outputs_path), "%s.outputs", selection_path);
-        if (!validate_scene_editor_outputs(ctx, obj_get(selection, "outputs"), outputs_path))
-            return false;
-        char hover_outputs_path[PATH_BUFFER_SIZE];
-        format_path(hover_outputs_path, sizeof(hover_outputs_path), "%s.hover_outputs", selection_path);
-        if (!validate_scene_editor_outputs(ctx, obj_get(selection, "hover_outputs"), hover_outputs_path))
-            return false;
-    }
-
-    char placement_path[PATH_BUFFER_SIZE];
-    format_path(placement_path, sizeof(placement_path), "%s.editor.placement", json_path);
-    if (!validate_scene_editor_placement(ctx, obj_get(editor, "placement"), placement_path, names))
-        return false;
-
-    yyjson_val *overlay = obj_get(editor, "debug_overlay");
-    if (overlay != NULL)
-    {
-        char overlay_path[PATH_BUFFER_SIZE];
-        format_path(overlay_path, sizeof(overlay_path), "%s.editor.debug_overlay", json_path);
-        if (!yyjson_is_obj(overlay))
-            return validation_error(ctx, overlay_path, "scene editor debug_overlay must be an object");
-        yyjson_val *enabled = obj_get(overlay, "enabled");
-        if (enabled != NULL && !yyjson_is_bool(enabled))
-            return validation_error(ctx, overlay_path, "scene editor debug_overlay enabled must be a boolean");
-        char flags_path[PATH_BUFFER_SIZE];
-        format_path(flags_path, sizeof(flags_path), "%s.flags", overlay_path);
-        if (!validate_string_or_string_array_names(ctx, obj_get(overlay, "flags"), flags_path, "editor debug flag",
-                                                   editor_debug_flag_name_valid))
-            return false;
-        static const char *const color_keys[] = {
-            "world_bounds_color", "selection_bounds_color", "trace_color",          "face_normal_color",
-            "hit_marker_color",   "command_preview_color",  "work_plane_grid_color"};
-        for (size_t i = 0; i < SDL_arraysize(color_keys); ++i)
-        {
-            yyjson_val *color = obj_get(overlay, color_keys[i]);
-            if (color != NULL && !is_exact_vec3_or_vec4_array(color))
-                return validation_error(ctx, overlay_path, "scene editor debug_overlay %s must be a vec3 or vec4 color",
-                                        color_keys[i]);
-        }
-        yyjson_val *normal_length = obj_get(overlay, "normal_length");
-        if (normal_length != NULL && (!yyjson_is_num(normal_length) || yyjson_get_num(normal_length) <= 0.0))
-            return validation_error(ctx, overlay_path, "scene editor debug_overlay normal_length must be positive");
-        yyjson_val *hit_marker_size = obj_get(overlay, "hit_marker_size");
-        if (hit_marker_size != NULL && (!yyjson_is_num(hit_marker_size) || yyjson_get_num(hit_marker_size) <= 0.0))
-            return validation_error(ctx, overlay_path, "scene editor debug_overlay hit_marker_size must be positive");
-        yyjson_val *grid_size = obj_get(overlay, "work_plane_grid_size");
-        if (grid_size != NULL && (!yyjson_is_num(grid_size) || yyjson_get_num(grid_size) <= 0.0))
-            return validation_error(ctx, overlay_path,
-                                    "scene editor debug_overlay work_plane_grid_size must be positive");
-        yyjson_val *grid_spacing = obj_get(overlay, "work_plane_grid_spacing");
-        if (grid_spacing != NULL && (!yyjson_is_num(grid_spacing) || yyjson_get_num(grid_spacing) <= 0.0))
-            return validation_error(ctx, overlay_path,
-                                    "scene editor debug_overlay work_plane_grid_spacing must be positive");
-    }
-    return true;
 }
 
 static bool validate_scene_details(validation_context *ctx, yyjson_val *root, yyjson_val *game_root,
