+++ src/game_data_validation_internal.h
@@ -95,11 +95,15 @@ bool require_unique_name(validation_context *ctx, name_table *table, const char 
                          const char *json_path);
 bool require_ref(validation_context *ctx, const name_table *table, const char *kind, const char *name,
                  const char *json_path);
+bool require_actor_ref(validation_context *ctx, const validation_names *names, const char *name, const char *json_path);
 bool asset_path_exists(validation_context *ctx, const char *asset_path, const char *json_path, const char *asset_kind);
 bool validate_editor_metadata(validation_context *ctx, yyjson_val *metadata, const char *json_path,
                               validation_names *names, bool allow_templates);
+bool validate_editor_metadata_tree(validation_context *ctx, yyjson_val *root, validation_names *names);
+bool validate_editor_player_starts(validation_context *ctx, yyjson_val *root, validation_names *names);
 bool require_unique_editor_stable_id(validation_context *ctx, name_table *stable_ids, yyjson_val *json,
                                      const char *json_path);
+bool validation_mouse_button_name_valid(const char *name);
 
 bool is_vec_array(yyjson_val *value, size_t min_count);
 bool is_exact_vec_array(yyjson_val *value, size_t count);
@@ -112,5 +116,7 @@ bool brush_surface_flag_name_valid(const char *name);
 bool validate_brush_string_or_string_array(validation_context *ctx, yyjson_val *value, const char *path,
                                            const char *label, bool (*name_valid)(const char *name), bool allow_empty);
 bool validate_brush_worlds(validation_context *ctx, yyjson_val *root, validation_names *names);
+bool validate_scene_editor_tooling(validation_context *ctx, yyjson_val *scene_root, const char *json_path,
+                                   validation_names *names);
 
 #endif /* SLAYER3D_GAME_DATA_VALIDATION_INTERNAL_H */
